Set .spec.startingDeadlineSeconds for group sync cronjobs
The Kubernetes Cronjob scheduler will refuse to create new jobs for a
cronjob which wasn't started for more than 100 times, from the [docs]:

> For every CronJob, the CronJob Controller checks how many schedules it
> missed in the duration from its last scheduled time until now. If there
> are more than 100 missed schedules, then it does not start the job and
> logs the error
>
> ```
> Cannot determine if job needs to be started. Too many missed start time (> 100). Set or decrease .spec.startingDeadlineSeconds or check clock skew.
> ```

However, if the `.spec.startingDeadlineSeconds` field is set, the
scheduler instead checks how many missed jobs occurred in the last
`startingDeadlineSeconds` seconds, instead of since the last schedule.

Therefore having `spec.startingDeadlineSeconds` set allows the cronjob
to be scheduled again after it didn't work for a while (for whatever
reason).

This commit sets `startingDeadlineSeconds` to 30 for the group sync
cronjob.

[docs]: https://kubernetes.io/docs/concepts/workloads/controllers/cron-jobs/