Make LDAP CA configuration optional
Previously, when configuring LDAP authentication with a server which is
secured by a certificate which is signed by a CA which is part of the
OCP4 system CA bundle (e.g. Let's Encrypt), users of the component still
had to supply the CA certificates in the configuration hierarchy.

With this commit, the component will ensure that the system CA bundle is
used for LDAP authentication configurations for which field `ca` is not
set.

This change enables users to omit CAs which are trusted through the
system CA bundle from their LDAP authentication configurations.

Please note that the LDAP group sync CronJob deployed by the component
is changed slightly to consume the LDAP server CA certificate from a
ConfigMap rather than the Secret containing the sync configuration. This
allows us to minimize the differences between configurations which use
the system CA bundle and configurations which provide a custom CA
certificate.

Resolves #27